Fixes
- Expand on error description
- Add private marker
- Remove TODO
- Remove obsolete snapshot
- Add `filter.ts`
- Other changes
