+++ packages/vitest/src/node/cli/cac.ts
@@ -1,11 +1,11 @@
 import type { VitestRunMode } from '../types/config'
+import type { CliOptions } from './cli-api'
 import type { CLIOption, CLIOptions as CLIOptionsConfig } from './cli-config'
 import { toArray } from '@vitest/utils'
 import cac, { type CAC, type Command } from 'cac'
 import { normalize } from 'pathe'
 import c from 'tinyrainbow'
 import { version } from '../../../package.json' with { type: 'json' }
-import { type CliOptions, outputFileList, processCollected } from './cli-api'
 import { benchCliOptionsConfig, cliOptionsConfig, collectCliOptionsConfig } from './cli-config'
 
 function addCommand(cli: CAC | Command, name: string, option: CLIOption<any>) {
@@ -300,7 +300,7 @@ async function collect(mode: VitestRunMode, cliFilters: string[], options: CliOp
   catch {}
 
   try {
-    const { prepareVitest } = await import('./cli-api')
+    const { prepareVitest, processCollected, outputFileList } = await import('./cli-api')
     const ctx = await prepareVitest(mode, {
       ...normalizeCliOptions(options),
       watch: false,
