+++ packages/vitest/src/node/core.ts
@@ -26,7 +26,7 @@ import { getCoverageProvider } from '../integrations/coverage'
 import { distDir } from '../paths'
 import { wildcardPatternToRegExp } from '../utils/base'
 import { VitestCache } from './cache'
-import { groupFilters, parseFilter } from './cli/cli-api'
+import { groupFilters, parseFilter } from './cli/filter'
 import { resolveConfig } from './config/resolveConfig'
 import { FilesNotFoundError, GitNotFoundError, IncludeTaskLocationDisabledError } from './errors'
 import { Logger } from './logger'
