+++ packages/vitest/src/typecheck/collect.ts
@@ -212,7 +212,6 @@ export async function collectTests(
   interpretTaskModes(
     file,
     ctx.config.testNamePattern,
-    // TODO: make sure you're not screwing stuff up
     undefined,
     hasOnly,
     false,
